move config package into runner
