+++ main.go
@@ -24,15 +24,16 @@ package main
 
 import (
 	"flag"
-	"github.com/Arriven/db1000n/src/metrics"
 	"log"
 	"os"
 	"os/signal"
 	"syscall"
 	"time"
 
-	"github.com/Arriven/db1000n/src/config"
+	"github.com/Arriven/db1000n/src/metrics"
+
 	"github.com/Arriven/db1000n/src/runner"
+	"github.com/Arriven/db1000n/src/runner/config"
 	"github.com/Arriven/db1000n/src/utils"
 	"github.com/Arriven/db1000n/src/utils/templates"
 )
